Fix empty result if process is altered
This patch updates the common command-handling code to
always retrieve the present Qemu VM. The internal gdb-pt-dump
state is reset if the PID of the qemu has changed.

This is added to handle the case when a qemu process is killed
and started, but the user wants to keep the gdb session active.
Previously, gdb-pt-dump would still cache the state of the older
qemu process.