Add custom kernels for golden image testing
This patch adds the infrastructure to generate custom
x86_64 kernels that enter long mode and a generate page
table. Each generated kernel binary generates a single
page table that is chosen at compile time.

For example, this patch adds these 3 tests for page table
generation:
- x86_64_test_entry_setup_4k_page_table_simple.bin
- x86_64_test_entry_setup_4k_page_table_complex.bin
- x86_64_test_entry_setup_2mb_page_table_simple.bin
Each binary is a floppy-disk image.

The binaries can be used to have the QEMU VM's page tables
in a deterministic state. The output can be used to generate
audited golden images that can be in the future compared
against when running integration tests.

Patch is for issue #11