pt: update x86_64 and x86_32 backend selection logic
This patch removes the x86_32 backend and unifies page table
parsing for x86_64 and x86_32. There is still some code that
branches off between 32-bit and 64-bit mode but this seems to
be cleaner than adding more abstraction for PML4 and PDP parsing.

The EFER long mode flag is chosen to select which architecture
should be used every time the page table is parsed. This
guarantees gdb_pt_dump always sees the current architectural state
of the VM.